Fix that member is logged out if his alias is changed

+++ src/main/java/de/altenerding/biber/pinkie/business/login/control/LoginModifier.java
@@ -21,32 +21,51 @@ public class LoginModifier {
 		logger.info("Setting new password for login for alias={}", alias);
 
 		Login login = loginProvider.getLoginByAlias(alias);
+
+        if (login == null) {
+            logger.error("No login credentials found for alias={}", alias);
+            return;
+        }
+
 		byte[] salt = Base64.getDecoder().decode(login.getSalt());
 		login.setSalt(Base64.getEncoder().encodeToString(salt));
 		login.setOnetimePassword(isOnetimePassword);
 		login.setLoginCount(0);
 		login.setPassword(securityProvider.hashPassword(passwordNew.toCharArray(), salt));
 
-		em.merge(login);
-		em.flush();
-	}
+        updateLogin(login);
+    }
 
-	public void updateAlias(String aliasOld, String aliasNew) throws Exception {
-		aliasOld = aliasOld.toLowerCase();
-		aliasNew = aliasNew.toLowerCase();
-		logger.info("Changing alias from {} to {}", aliasOld, aliasNew);
-		Login login = loginProvider.getLoginByAlias(aliasOld);
+    public void updateAlias(String aliasOld, String aliasNew) throws Exception {
+        aliasOld = aliasOld.toLowerCase();
+        aliasNew = aliasNew.toLowerCase();
+        logger.info("Changing alias from {} to {}", aliasOld, aliasNew);
+        Login login = loginProvider.getLoginByAlias(aliasOld);
 
-		login.setAlias(aliasNew);
-		em.merge(login);
-		em.flush();
-		logger.info("Successfully changed alias from {} to {}", aliasOld, aliasNew);
-	}
+        if (login == null) {
+            logger.error("No login credentials found for old alias={}", aliasOld);
+            return;
+        }
+
+        login.setAlias(aliasNew);
+        updateLogin(login);
+        logger.info("Successfully changed alias from {} to {}", aliasOld, aliasNew);
+    }
+
+    void updateLogin(Login login) {
+        em.merge(login);
+        em.flush();
+    }
 
 	public void removeLoginForAlias(String alias) throws Exception {
 		alias = alias.toLowerCase();
-		logger.info("Removing login fpr alias={}", alias);
+        logger.info("Removing login for alias={}", alias);
 		Login login = loginProvider.getLoginByAlias(alias);
+
+        if (login == null) {
+            logger.error("No login credentials found for alias={}", alias);
+            return;
+        }
 
 		login = em.merge(login);
 		em.remove(login);
